cr: removed logic.go
