+++ api/api_client.go
@@ -38,3 +38,15 @@ func (self *ApiClient) normalizeEndpoint() {
 		self.Endpoint = self.Endpoint[:len(self.Endpoint)-1]
 	}
 }
+
+func (self *ApiClient) organizationId() (string, error) {
+	if self.cachedOrganizationId != "" {
+		return self.cachedOrganizationId, nil
+	}
+	organization, err := self.Organization()
+	if err != nil {
+		return "", nil
+	}
+	self.cachedOrganizationId = organization.Id
+	return self.cachedOrganizationId, nil
+}
