--- api/logic.go
@@ -1,13 +1,0 @@
-package api
-
-func (self *ApiClient) organizationId() (string, error) {
-	if self.cachedOrganizationId != "" {
-		return self.cachedOrganizationId, nil
-	}
-	organization, err := self.Organization()
-	if err != nil {
-		return "", nil
-	}
-	self.cachedOrganizationId = organization.Id
-	return self.cachedOrganizationId, nil
-}
